% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_presentation.R
\name{create_presentation}
\alias{create_presentation}
\title{Create a new presentation}
\usage{
create_presentation(file_name, title, authors, subtitle = "")
}
\arguments{
\item{file_name}{The name of the file to create (e.g., "presentation.qmd"). If it does not end in .qmd, the extension will be appended.}

\item{title}{The title of the presentation.}

\item{authors}{A character vector of author names.}

\item{subtitle}{The subtitle of the presentation (optional).}
}
\value{
The name of the created file (invisibly).
}
\description{
This function creates a new .qmd file for a RevealJS presentation using the Baylor Statistics Consulting Center theme.
}
