% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a new report}
\usage{
create_report(
  file_name,
  title,
  client,
  authors,
  email = "rodney_strudivant@baylor.edu",
  phone = "(254) 710-1663",
  keywords = NULL
)
}
\arguments{
\item{file_name}{The name of the file to create (e.g., "analysis.qmd"). If it does not end in .qmd, the extension will be appended.}

\item{title}{The title of the project.}

\item{client}{The name of the client.}

\item{authors}{A character vector of author names.}

\item{email}{The email address for the contact info (default: "rodney_strudivant@baylor.edu").}

\item{phone}{The phone number for the contact info (default: "(254) 710-1663").}

\item{keywords}{Optional keywords to include in the report.}
}
\value{
The name of the created file (invisibly).
}
\description{
This function creates a new .qmd file with a standardized YAML header for Baylor Statistics Consulting Center projects.
}
\examples{
\dontrun{
create_report("project_analysis", "New Project", "Dr. Client", c("Student A", "Student B"))
}
}
