% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_path.R
\name{get_box_root}
\alias{get_box_root}
\title{Get the configured Box root folder}
\usage{
get_box_root()
}
\value{
A character string containing the normalized Box root path.
}
\description{
Returns the Box root directory previously set with
\code{set_box_root()}.
}
